requirements.txt
Flask==1.1.2
requests==2.25.1
flask_restful==0.3.9
python-dotenv
celery==5.1.0
Jinja2==3.0.1
itsdangerous==2.0.1
Werkzeug==2.0.1
marshmallow-sqlalchemy==0.25.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.15
redis==3.5.3